# docs/api/openAPI.yaml
components:
  schemas:
...
    AddressDto:
      additionalProperties: false
      properties:
        bpna:
          maxItems: 50
          pattern: ^BPNA[0-9a-zA-Z]{12}$
          type: string
        country:
          maxItems: 50
          pattern: ^[^\n\x0B\f\r\x85\u2028\u2029]+$
          type: string
        streetAndNumber:
          maxItems: 50
          pattern: ^[^\n\x0B\f\r\x85\u2028\u2029]+$
          type: string
        zipCodeAndCity:
          maxItems: 50
          pattern: ^[^\n\x0B\f\r\x85\u2028\u2029]+$
          type: string
      type: object
...
    Classification:
      additionalProperties: false
      properties:
        classificationDescription:
          maxItems: 50
          pattern: ^[^\n\x0B\f\r\x85\u2028\u2029]+$
          type: string
        classificationID:
          maxItems: 50
          pattern: ^[^\n\x0B\f\r\x85\u2028\u2029]+$
          type: string
        classificationStandard:
          maxItems: 50
          pattern: ^[^\n\x0B\f\r\x85\u2028\u2029]+$
          type: string
      required:
      - classificationID
      - classificationStandard
      type: object
...
    MaterialEntityDto:
      additionalProperties: false
      properties:
        materialFlag:
          type: boolean
        materialNumberCx:
          maxItems: 50
          pattern: (^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$)|(^urn:uuid:[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$)
          type: string
        name:
          maxItems: 50
          pattern: ^[^\n\x0B\f\r\x85\u2028\u2029]+$
          type: string
        ownMaterialNumber:
          maxItems: 50
          pattern: ^[^\n\x0B\f\r\x85\u2028\u2029]+$
          type: string
        productFlag:
          type: boolean
      type: object
    MaterialPartnerRelationDto:
      additionalProperties: false
      properties:
        ownMaterialNumber:
          maxItems: 50
          pattern: ^[^\n\x0B\f\r\x85\u2028\u2029]+$
          type: string
        ownProducingSiteBpnss:
          items:
            maxItems: 50
            pattern: ^BPNS[0-9a-zA-Z]{12}$
            type: string
          maxItems: 50
          type: array
        ownStockingSiteBpnss:
          items:
            maxItems: 50
            pattern: ^BPNS[0-9a-zA-Z]{12}$
            type: string
          maxItems: 50
          type: array
        partnerBpnl:
          maxItems: 50
          pattern: ^BPNL[0-9a-zA-Z]{12}$
          type: string
        partnerBuysMaterial:
          type: boolean
        partnerMaterialNumber:
          maxItems: 50
          pattern: ^[^\n\x0B\f\r\x85\u2028\u2029]+$
          type: string
        partnerSuppliesMaterial:
          type: boolean
      type: object
...
    PartSitesInformationAsPlanned:
      additionalProperties: false
      properties:
        catenaXsiteId:
          maxItems: 50
          pattern: ^BPNS[0-9a-zA-Z]{12}$
          type: string
        function:
          enum:
          - production
          - warehouse
          - spare part warehouse
          maxItems: 50
          type: string
        functionValidFrom:
          maxItems: 50
          pattern: ^(?:[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}(?:[.][0-9]+)?Z|[0-9]{4}-[0-9]{2}-[0-9]{2}(?:T[0-9]{2}:[0-9]{2}:[0-9]{2}(?:[.][0-9]+)?(?:Z|[+-][0-9]{2}:[0-9]{2}))?)$
          type: string
        functionValidUntil:
          maxItems: 50
          pattern: ^(?:[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}(?:[.][0-9]+)?Z|[0-9]{4}-[0-9]{2}-[0-9]{2}(?:T[0-9]{2}:[0-9]{2}:[0-9]{2}(?:[.][0-9]+)?(?:Z|[+-][0-9]{2}:[0-9]{2}))?)$
          type: string
      required:
      - catenaXsiteId
      - function
      type: object
    PartTypeInformationBody:
      additionalProperties: false
      properties:
        manufacturerPartId:
          maxItems: 50
          type: string
        nameAtManufacturer:
          maxItems: 50
          type: string
        partClassification:
          items:
            $ref: '#/components/schemas/Classification'
          maxItems: 50
          type: array
      required:
      - manufacturerPartId
      - nameAtManufacturer
      type: object
    PartTypeInformationSAMM:
      additionalProperties: false
      properties:
        catenaXId:
          maxItems: 50
          pattern: (^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$)|(^urn:uuid:[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$)
          type: string
        partSitesInformationAsPlanned:
          items:
            $ref: '#/components/schemas/PartSitesInformationAsPlanned'
          maxItems: 50
          type: array
          uniqueItems: true
        partTypeInformation:
          $ref: '#/components/schemas/PartTypeInformationBody'
      required:
      - catenaXId
      - partTypeInformation
      type: object
...
    PartnerDto:
      additionalProperties: false
      properties:
        addresses:
          items:
            $ref: '#/components/schemas/AddressDto'
          maxItems: 50
          type: array
          uniqueItems: true
        bpnl:
          maxItems: 50
          pattern: ^BPNL[0-9a-zA-Z]{12}$
          type: string
        edcUrl:
          maxItems: 50
          pattern: ^http[s]?://([a-z0-9][a-z0-9\-]+[a-z0-9])(\.[a-z0-9\-]+)*(:[0-9]{1,4})?(/[a-z0-9\-]+)*[/]?$
          type: string
        name:
          maxItems: 50
          pattern: ^[^\n\x0B\f\r\x85\u2028\u2029]+$
          type: string
        sites:
          items:
            $ref: '#/components/schemas/SiteDto'
          maxItems: 50
          type: array
          uniqueItems: true
        uuid:
          format: uuid
          maxItems: 50
          type: string
      type: object
...
    SiteDto:
      additionalProperties: false
      properties:
        addresses:
          items:
            $ref: '#/components/schemas/AddressDto'
          maxItems: 50
          type: array
          uniqueItems: true
        bpns:
          maxItems: 50
          pattern: ^BPNS[0-9a-zA-Z]{12}$
          type: string
        name:
          maxItems: 50
          pattern: ^[^\n\x0B\f\r\x85\u2028\u2029]+$
          type: string
      type: object
...
  securitySchemes:
    X-API-KEY:
      in: header
      name: X-API-KEY
      type: apiKey
info:
  title: PURIS FOSS Open API
  version: 6.0.0
openapi: 3.1.0
paths:
...
  /materialpartnerrelations:
    get:
      description: Returns a list of all MaterialPartnerRelations as DTOs.
      operationId: getAllMaterialPartnerRelations
      responses:
        '200':
          content:
            '*/*':
              schema:
                additionalProperties: false
                items:
                  $ref: '#/components/schemas/MaterialPartnerRelationDto'
                maxItems: 50
                type: array
          description: Successfully retrieved all MaterialPartnerRelations.
        '500':
          content:
            '*/*':
              schema:
                additionalProperties: false
                items:
                  $ref: '#/components/schemas/MaterialPartnerRelationDto'
                maxItems: 50
                type: array
          description: Internal Server Error.
      summary: Get all Material Partner Relations -- ADMIN ONLY
      tags:
      - material-partner-relations-controller
    post:
      description: 'Creates a new MaterialPartnerRelation with the given parameter
        data. Please note that this is only possible, if the designated Material and
        Partner entities have already been created before this request. '
      operationId: createMaterialPartnerRelation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialPartnerRelationDto'
              additionalProperties: false
        required: true
      responses:
        '200':
          content:
            '*/*':
              schema:
                additionalProperties: false
                type: object
          description: Successfully created a new MaterialPartnerRelationEntity.
        '400':
          content:
            '*/*':
              schema:
                additionalProperties: false
                type: object
          description: Material and/or Partner do not exist or invalid parameters
        '409':
          content:
            '*/*':
              schema:
                additionalProperties: false
                type: object
          description: Relation for given Material and Partner does already exist.
        '500':
          content:
            '*/*':
              schema:
                additionalProperties: false
                type: object
          description: Internal Server Error.
      summary: Creates a Material Partner Relation -- ADMIN ONLY
      tags:
      - material-partner-relations-controller
    put:
      description: 'Updates an existing MaterialPartnerRelation. You have to specify
        the ownMaterialNumber and the partnerBpnl. The other three parameters are
        genuinely optional. Provide them only if you want to change their values. '
      operationId: updateMaterialPartnerRelation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialPartnerRelationDto'
              additionalProperties: false
        required: true
      responses:
        '204':
          content:
            '*/*':
              schema:
                additionalProperties: false
                type: object
          description: Update was accepted.
        '400':
          content:
            '*/*':
              schema:
                additionalProperties: false
                type: object
          description: Invalid parameters
        '404':
          content:
            '*/*':
              schema:
                additionalProperties: false
                type: object
          description: No existing entity was found.
        '500':
          content:
            '*/*':
              schema:
                additionalProperties: false
                type: object
          description: Internal Server Error.
      summary: Updates a Material Partner Relation -- ADMIN ONLY
      tags:
      - material-partner-relations-controller
  /materials:
    get:
      description: Returns the requested Material dto, specified by the given ownMaterialNumber.
      operationId: getMaterial
      parameters:
      - description: The Material Number that is used in your own company to identify
          the Material, encoded in base64
        in: query
        name: ownMaterialNumber
        required: true
        schema:
          additionalProperties: false
          maxItems: 50
          type: string
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MaterialEntityDto'
                additionalProperties: false
          description: Returns the requested Material.
        '400':
          description: Invalid parameter
        '404':
          description: Requested Material was not found.
      summary: Gets a Material by ownMaterialNumber
      tags:
      - material-controller
    post:
      description: Creates a new Material entity with the data given in the request
        body. As a bare minimum, it must contain a new, unique ownMaterialNumber.
      operationId: createMaterial
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialEntityDto'
              additionalProperties: false
        required: true
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MaterialEntityDto'
                additionalProperties: false
          description: Successfully created a new Material entity.
        '400':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MaterialEntityDto'
                additionalProperties: false
          description: Malformed request body.
        '409':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MaterialEntityDto'
                additionalProperties: false
          description: Material with the given ownMaterialNumber already exists.
        '500':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MaterialEntityDto'
                additionalProperties: false
          description: Internal Server error.
      summary: Creates a Material -- ADMIN ONLY
      tags:
      - material-controller
    put:
      description: Updates an existing Material entity with the data given in the
        request body.
      operationId: updateMaterial
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialEntityDto'
              additionalProperties: false
        required: true
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MaterialEntityDto'
                additionalProperties: false
          description: Update was accepted.
        '400':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MaterialEntityDto'
                additionalProperties: false
          description: Malformed request body.
        '404':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MaterialEntityDto'
                additionalProperties: false
          description: No existing Material Entity found, no update was performed.
        '500':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MaterialEntityDto'
                additionalProperties: false
          description: Internal Server Error.
      summary: Updates a Material -- ADMIN ONLY
      tags:
      - material-controller
...
  /partners:
    get:
      description: Returns the requested PartnerDto.
      operationId: getPartner
      parameters:
      - description: The unique BPNL that was assigned to that Partner.
        example: BPNL2222222222RR
        in: query
        name: partnerBpnl
        required: true
        schema:
          additionalProperties: false
          maxItems: 50
          type: string
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PartnerDto'
                additionalProperties: false
          description: Found Partner, returning it in response body.
        '400':
          description: Invalid parameter.
        '404':
          description: Requested Partner not found.
        '500':
          description: Internal Server Error.
      summary: Gets a specific Partner -- ADMIN ONLY
      tags:
      - partner-controller
    post:
      description: Creates a new Partner entity with the data given in the request
        body. Please note that no UUID can be assigned to a Partner that wasn't created
        before. So the request body must not contain a UUID.
      operationId: createPartner
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartnerDto'
              additionalProperties: false
        required: true
      responses:
        '200':
          content:
            '*/*':
              schema:
                additionalProperties: false
                type: object
          description: Partner created successfully.
        '400':
          content:
            '*/*':
              schema:
                additionalProperties: false
                type: object
          description: Request body was malformed, didn't meet the minimum constraints
            or wrongfully contained a UUID.
        '409':
          content:
            '*/*':
              schema:
                additionalProperties: false
                type: object
          description: 'The BPNL specified in the request body is already assigned. '
      summary: Creates a new Partner -- ADMIN ONLY
      tags:
      - partner-controller
...
  /parttypeinformation/{globalAssetId}/submodel/{representation}:
    get:
      description: 'Endpoint that delivers PartTypeInformation of own products to
        customer partners. ''globalAssetId'' must be set to the global asset id of
        the party, that receives the request. Please note that the SAMMs delivered
        by this endpoint don''t provide partClassification data. This endpoint is
        meant to be accessed by partners via EDC only. '
      operationId: getMapping
      parameters:
      - in: header
        name: edc-bpn
        required: true
        schema:
          additionalProperties: false
          maxItems: 50
          type: string
      - description: The global asset id for the requested part type.
        in: path
        name: globalAssetId
        required: true
        schema:
          additionalProperties: false
          maxItems: 50
          type: string
      - description: Must be set to '$value'
        in: path
        name: representation
        required: true
        schema:
          additionalProperties: false
          maxItems: 50
          type: string
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PartTypeInformationSAMM'
                additionalProperties: false
          description: Ok
        '400':
          description: 'Invalid request parameters. '
        '401':
          description: 'Access forbidden. '
        '404':
          description: 'Product not found for given parameters. '
        '501':
          description: 'Unsupported representation requested. '
      tags:
      - part-type-information-controller
...
security:
- X-API-KEY: []
servers:
- description: Generated server url
  url: http://localhost:8081/catena